Pulled out the entire ui configuration in ui_client
